Sorry @kureev, your prefix idea was better :muscle:
